refactor: 부분 Suspense 사용 및 Navigation 최적화
