--- src/app/(navigation)/home/loading.tsx
@@ -1,57 +1,0 @@
-export default function Loading() {
-  return (
-    <div className="flex flex-col gap-10">
-      <div className="animate-pulse flex flex-col gap-10">
-        <div className="h-[22px] w-full shrink-0 rounded-full bg-gray-500" />
-        <div className="h-[358px] w-full shrink-0 rounded-lg bg-gray-500" />
-      </div>
-      <div className="animate-pulse flex flex-col gap-10">
-        <div className="h-[22px] w-full shrink-0 rounded-full bg-gray-500" />
-        <div className="flex gap-2">
-          <div className="h-[128px] w-[122px] shrink-0 rounded-lg bg-gray-500" />
-          <div className="flex flex-col gap-5 w-full justify-center">
-            <div className="h-[19px] w-full rounded-full bg-gray-500" />
-            <div className="h-[19px] w-full  rounded-full bg-gray-500" />
-            <div className="flex justify-between gap-14">
-              <div className="h-[19px] w-full rounded-full bg-gray-500" />
-              <div className="h-[19px] w-full rounded-full bg-gray-500" />
-            </div>
-          </div>
-        </div>
-        <div className="flex gap-2">
-          <div className="h-[128px] w-[122px] shrink-0 rounded-lg bg-gray-500" />
-          <div className="flex flex-col gap-5 w-full justify-center">
-            <div className="h-[19px] w-full rounded-full bg-gray-500" />
-            <div className="h-[19px] w-full  rounded-full bg-gray-500" />
-            <div className="flex justify-between gap-14">
-              <div className="h-[19px] w-full rounded-full bg-gray-500" />
-              <div className="h-[19px] w-full rounded-full bg-gray-500" />
-            </div>
-          </div>
-        </div>
-        <div className="flex gap-2">
-          <div className="h-[128px] w-[122px] shrink-0 rounded-lg bg-gray-500" />
-          <div className="flex flex-col gap-5 w-full justify-center">
-            <div className="h-[19px] w-full rounded-full bg-gray-500" />
-            <div className="h-[19px] w-full  rounded-full bg-gray-500" />
-            <div className="flex justify-between gap-14">
-              <div className="h-[19px] w-full rounded-full bg-gray-500" />
-              <div className="h-[19px] w-full rounded-full bg-gray-500" />
-            </div>
-          </div>
-        </div>
-        <div className="flex gap-2">
-          <div className="h-[128px] w-[122px] shrink-0 rounded-lg bg-gray-500" />
-          <div className="flex flex-col gap-5 w-full justify-center">
-            <div className="h-[19px] w-full rounded-full bg-gray-500" />
-            <div className="h-[19px] w-full  rounded-full bg-gray-500" />
-            <div className="flex justify-between gap-14">
-              <div className="h-[19px] w-full rounded-full bg-gray-500" />
-              <div className="h-[19px] w-full rounded-full bg-gray-500" />
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
